requirements.txt
Flask==2.3.3
requests==2.31.0
asyncio==3.4.3
aiohttp==3.11.18
# polars==1.27.1
google-cloud-storage==3.1.0
google-cloud-bigquery==3.31.0
google-cloud-logging==3.12.1
pytest==8.3.5
python-dotenv==1.1.0
debugpy # Required for debugging.
